Added XDG gem.
Provides an implementation of the XDG Base Directory Specification
which helps CLIs and other programs leverage a common structure for
configuration, cache, and data.

In the past, this gem wasn't well maintained. Prior to Version 3.0.0,
it hadn't been updated in ~7 years. Now, with new ownership, it's got a
new implementation and API which adheres to the XDG Base Directory
Specification.

Further enhancements and examples can be found in the
[Runcom](https://github.com/bkuhlmann/runcom) gem.