Add ComfyBootstrapForm (Form Builder) (#799)
Currently, this is the only form builder that properly supports Bootstrap 4 markup and works in Rails 5.2 by wrapping`form_with` form helper. Rails 5.0+ versions are also supported as `form_for` is wrapped as well. Other form builders are either very out-of-date or not planning to support Bootstrap 4 at all.

Now it's a direct dependency of [ComfortableMexicanSofa CMS](https://github.com/comfy/comfortable-mexican-sofa) that uses Bootstrap 4 in the admin area.