Add Cloud 66
Add Cloud 66 to DevOps category.
- Website: https://www.cloud66.com/
- Help page: https://help.cloud66.com/

What is this Ruby project?
Cloud 66 - DevOps as a service that helps to build, deploy and manage Ruby application on any cloud or server.
What are the main difference between this Ruby project and similar ones?
Cloud 66 provides services like a PaaS platform but on your own cloud or server.